Add GetVolunteerCoordinatesRequest ToString() tests (because it's used as a cache key).  Move GetVolunteerCoordinatesRequest class out of Handlers project.
