+++ UserService.UnitTests/GetVolunteerCoordinatesHandlerTests.cs
@@ -5,6 +5,7 @@ using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 using UserService.Core;
+using UserService.Core.BusinessLogic;
 using UserService.Core.Cache;
 using UserService.Core.Domains.Entities;
 using UserService.Handlers;
